CorrectionMap now two vectors instead of a vector of pairs.  This was because I thought
root was geting confused by it.  Not sure if that was a problem but for I will leave it.
The memory dump issue when reading events back in from file was because the I was calling define map in
cosntructor.  Root didn't like that for some reason.

things seem to be working.  need to do some clean up and optimization