Added vector support to the Correctable class.  Now has two main maps. theVariableMap which
holds references to simple variable types.  It is with void pointers to all for any simple type int, double...

the other map is the theVectorVariableMap which goes from strings to vector<Double_t>* so onlt that type is allowed.

For the vectors to work.  Entries in both maps is needed.  To access spots in "energies" it must be added to both:
theVariableMap["energies"]=&energies
theVectorVariableMap["energies"]=&energies

it is need in the normal map in order for checks of Get("  ") to work.  It is needed in the other one so you don't need to cast vectors of
things from void pointers.  This appears to work